+++ custom_components/s2jyarbo/__init__.py
@@ -11,7 +11,7 @@ from .const import DOMAIN
 from .mqtt import YarboMqttClient
 from .panel import async_register_panel
 
-PLATFORMS: list[Platform] = [Platform.SENSOR, Platform.DEVICE_TRACKER]
+PLATFORMS: list[Platform] = [Platform.SENSOR, Platform.DEVICE_TRACKER, Platform.EVENT]
 CONFIG_SCHEMA = cv.config_entry_only_config_schema(DOMAIN)